// packages/shop-state-management/src/modules/products/actions.types.ts
export interface CreateProductRequest extends Action {
  type: typeof Constants.CREATE_PRODUCT
  payload: TYPES.MODELS.PRODUCTS.IProduct
}

export interface CreateProductSuccess extends Action {
  type: typeof Constants.CREATE_PRODUCT_SUCCESS
}

export interface CreateProductFailed extends Action {
  type: typeof Constants.CREATE_PRODUCT_FAILED
  payload: string
}

export interface DeleteProductRequest extends Action {
  type: typeof Constants.DELETE_PRODUCT
  payload: string
}
// ...
export interface TrashRequestList extends Action {
  type: typeof Constants.TRASH_PRODUCT_LIST
  payload: string
}

export interface TrashRequestSuccess {
  type: typeof Constants.TRASH_PRODUCT_LIST_SUCCESS
  payload: TYPES.MODELS.PRODUCTS.IResponseProductList
}

export interface TrashRequestFailed {
  type: typeof Constants.TRASH_PRODUCT_LIST_FAILED
  payload: string
}

export interface RestoreProduct extends Action {
  type: typeof Constants.RESTORE_PRODUCT
  payload: string
}